% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Batting.R
\docType{data}
\name{Batting}
\alias{Batting}
\title{Batting data}
\format{
A data frame with 112184 rows and 22 columns:
\describe{
\item{\code{playerID}}{\verb{[character]} Lahman ID for MLB Player}
}
}
\usage{
Batting
}
\description{
TODO: add description and more
}
\keyword{datasets}
