% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistics.R
\name{calcBA}
\alias{calcBA}
\title{Calculate batting average and add to data frame}
\usage{
calcBA(data)
}
\arguments{
\item{data}{A data frame with columns H (hits) and AB (at-bats)}
}
\value{
The input data frame with added BA column
}
\description{
Calculate batting average and add to data frame
}
\examples{
calcBA(Batting)
}
